# openapi.yaml
openapi: 3.0.3
info:
  title: Etranzact Test API
  version: 1.0.0
  description: OpenAPI spec generated from the current Spring controllers.
servers:
  - url: http://localhost:8080

tags:
  - name: Customers
  - name: Accounts
  - name: Transactions
  - name: Demo

paths:
  /api/v1/customers/signup:
    post:
      tags: [Customers]
      summary: Create a new customer (and primary account)
      operationId: customerSignup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
            example:
              firstName: Jane
              lastName: Doe
              dob: "1997-04-23"
              idNumber: "A12345678"
              pin: "1234"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignupResponse"
              example:
                customerId: 1
                primaryAccountId: 10
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"

  /api/v1/customers/login:
    post:
      tags: [Customers]
      summary: Login customer and receive auth token
      operationId: customerLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerLoginRequest"
            example:
              idNumber: "A12345678"
              pin: "1234"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
              example:
                token: "token-value"
                customerId: 1
                expiresAt: "2026-05-05T10:07:00Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  /api/v1/accounts/{accountId}/deposit:
    post:
      tags: [Accounts]
      summary: Deposit into an account
      operationId: deposit
      parameters:
        - $ref: "#/components/parameters/AccountId"
        - $ref: "#/components/parameters/AuthToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AmountRequest"
            example:
              amount: 100.00
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionMutationResponse"
              example:
                transactionId: 123
                balance: 250.00
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"

  /api/v1/accounts/{accountId}/withdraw:
    post:
      tags: [Accounts]
      summary: Withdraw from an account
      operationId: withdraw
      parameters:
        - $ref: "#/components/parameters/AccountId"
        - $ref: "#/components/parameters/AuthToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AmountRequest"
            example:
              amount: 50.00
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionMutationResponse"
              example:
                transactionId: 124
                balance: 200.00
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"

  /api/v1/accounts/{accountId}/transactions:
    get:
      tags: [Transactions]
      summary: Get paginated transaction history for an account
      operationId: transactionHistory
      parameters:
        - $ref: "#/components/parameters/AccountId"
        - $ref: "#/components/parameters/AuthToken"
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionHistoryPageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  /api/v1/login:
    post:
      tags: [Demo]
      summary: Demo login (echoes request body)
      operationId: demoLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequestDTO"
            example:
              username: demo
              password: secret
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginRequestDTO"

  /api/v1/populate-db:
    post:
      tags: [Demo]
      summary: Demo call to external API and return raw JSON string
      operationId: populateDb
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                type: string
            application/json:
              schema:
                type: string

  /api/v1/get-gb:
    get:
      tags: [Demo]
      summary: Demo health-ish endpoint
      operationId: getGb
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                type: string
              example: Get db api working

  /api/v1/test:
    get:
      tags: [Demo]
      summary: Demo endpoint returning a DTO (GET with a request body)
      operationId: demoTest
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequestDTO"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DemoResponseDto"
              example:
                value: Response from DTO

  /api/v1/test-with-service:
    get:
      tags: [Demo]
      summary: Demo endpoint using a service
      operationId: demoWithService
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                type: string

components:
  parameters:
    AccountId:
      name: accountId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    AuthToken:
      name: X-Auth-Token
      in: header
      required: true
      description: Token returned by `POST /api/v1/customers/login`.
      schema:
        type: string

  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"

  schemas:
    SignupRequest:
      type: object
      required: [firstName, lastName, dob, idNumber, pin]
      properties:
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        dob:
          type: string
          format: date
        idNumber:
          type: string
          minLength: 1
        pin:
          type: string
          minLength: 1

    SignupResponse:
      type: object
      required: [customerId, primaryAccountId]
      properties:
        customerId:
          type: integer
          format: int64
        primaryAccountId:
          type: integer
          format: int64

    CustomerLoginRequest:
      type: object
      required: [idNumber, pin]
      properties:
        idNumber:
          type: string
          minLength: 1
        pin:
          type: string
          minLength: 1

    LoginResponse:
      type: object
      required: [token, customerId, expiresAt]
      properties:
        token:
          type: string
        customerId:
          type: integer
          format: int64
        expiresAt:
          type: string
          format: date-time

    AmountRequest:
      type: object
      required: [amount]
      properties:
        amount:
          type: number
          format: double
          minimum: 0.0000000001

    TransactionMutationResponse:
      type: object
      required: [transactionId, balance]
      properties:
        transactionId:
          type: integer
          format: int64
        balance:
          type: number
          format: double

    TransactionType:
      type: string
      enum: [DEPOSIT, WITHDRAW]

    TransactionHistoryItem:
      type: object
      required: [id, type, amount, balanceBefore, balanceAfter, occurredAt]
      properties:
        id:
          type: integer
          format: int64
        type:
          $ref: "#/components/schemas/TransactionType"
        amount:
          type: number
          format: double
        balanceBefore:
          type: number
          format: double
        balanceAfter:
          type: number
          format: double
        occurredAt:
          type: string
          format: date-time

    TransactionHistoryPageResponse:
      type: object
      required: [content, totalPages, totalElements, number, size]
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/TransactionHistoryItem"
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        number:
          type: integer
          format: int32
        size:
          type: integer
          format: int32

    LoginRequestDTO:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string

    DemoResponseDto:
      type: object
      required: [value]
      properties:
        value:
          type: string

    ApiError:
      type: object
      required: [timestamp, status, error, message]
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        error:
          type: string
        message:
          type: string
        fieldErrors:
          type: object
          additionalProperties:
            type: string