# src/main/java/com/gmfs/demo/dto/LoginResponse.java
package com.gmfs.demo.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.time.Instant;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class LoginResponse {

    private String token;
    private Long customerId;
    private Instant expiresAt;
}