# src/main/java/com/gmfs/demo/dto/SignupRequest.java
package com.gmfs.demo.dto;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Past;
import lombok.Data;

import java.time.LocalDate;

@Data
public class SignupRequest {

    @NotBlank
    private String firstName;

    @NotBlank
    private String lastName;

    @NotNull
    @Past
    private LocalDate dob;

    @NotBlank
    private String idNumber;

    @NotBlank
    private String pin;
}