# src/main/java/com/gmfs/demo/dto/SignupResponse.java
package com.gmfs.demo.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class SignupResponse {

    private Long customerId;
    private Long primaryAccountId;
}