# src/main/java/com/gmfs/demo/dto/TransactionHistoryItem.java
package com.gmfs.demo.dto;

import com.gmfs.demo.model.TransactionType;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.math.BigDecimal;
import java.time.Instant;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class TransactionHistoryItem {

    private Long id;
    private TransactionType type;
    private BigDecimal amount;
    private BigDecimal balanceBefore;
    private BigDecimal balanceAfter;
    private Instant occurredAt;
}