# src/main/java/com/gmfs/demo/dto/TransactionMutationResponse.java
package com.gmfs.demo.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.math.BigDecimal;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class TransactionMutationResponse {

    private Long transactionId;
    private BigDecimal balance;
}